+++ BlueprintPurge/changelog.md
@@ -1,8 +1,10 @@
 ﻿# Changelog
 
 ## [0.2.0]
-- fixed missing comma with null replacement
 - fixed rare case where a wrong id would be parsed, even tho it doesn't have any id
+- changed so affected list are cleared instead of setting them to null; renamed 'Null' field to 'IsList'
+- added right click event to copy fields into clipboard
+- added 'Generate Peeks' option to reduce RAM usage, also faster when processing lots of guids
 
 ## [0.1.0]
 - first build